+++ apis/r/src/groups.cpp
@@ -153,3 +153,11 @@ void c_group_put_metadata(Rcpp::XPtr<somaobj_wrap_t> xp, std::string key, SEXP o
 
     soma_set_metadata(xp, key, obj);
 }
+
+// [[Rcpp::export]]
+bool soma_group_member_is_relative(Rcpp::XPtr<tiledbsoma::SOMAGroup> group, const std::string& name) {
+    if (!group) {
+        throw Rcpp::exception("Internal error: SOMAObject handle is not initialized.");
+    }
+    return group->is_relative(name);
+}