+++ apis/python/HISTORY.md
@@ -24,6 +24,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Removed
 
+- \[[#4431](https://github.com/single-cell-data/TileDB-SOMA/pull/4431)\] Remove deprecated support for allowing a dimension in `shape` to be `None` in the `SparseNDArray`.
+- \[[#4431](https://github.com/single-cell-data/TileDB-SOMA/pull/4431)\] Remove deprecated support for setting `domain=None` in the `create` method for `DataFrame`, `PointCloudDataFrame`, and `GeometryDataFrame`.
+
 ### Fixed
 
 ### Security